# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: hyve API
  description: |
    hyve's HTTP API + auth layer — see HYVE-CONTROLLER-ARCHITECTURE-PLAN.md's
    Phase 6. A thin, authorized front door onto the ClusterDefinition/
    Template/Workflow/Module CRDs the controller already reconciles —
    cluster mode never requires this API, plain kubectl against the CRDs
    always works. Organization/environment/RBAC-binding/session data lives
    in a separate Postgres/SQLite store (internal/orgdb), not Kubernetes —
    see HYVE-ORGANIZATION-MODEL-IMPLEMENTATION-PLAN.md, nexus-config/docs.

    Everything under /api/* requires a session: log in via /auth/login,
    then send the returned accessToken as `Authorization: Bearer
    <accessToken>` on every subsequent request. accessToken is short-lived
    (see AccessTokenTTL) — exchange sessionToken for a new one via
    /auth/refresh as needed, without re-authenticating with a
    username/password. A session is revocable server-side (backed by a
    Session row) via /auth/logout — unlike accessToken, which is a
    stateless signed value that can't be invalidated early.
  version: "0.1.0"
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/clusters/{name}/auth-context:
    get:
      tags: [clusters]
      summary: Get the data needed to run a driver module's auth op client-side
      description: |
        Only serves clusters using the default client-side auth method
        (spec.access.method unset) — a cluster using the module-auth or
        tunnel access override, or the API's own configured primary
        cluster, is rejected with 409 (fetch its kubeconfig via
        GET /api/kubeconfig instead). Deliberately more sensitive than
        ClusterDTO: this is the one legitimate place driverOutputs/params
        get exposed, since a client-side `hyve cluster auth` needs them to
        actually run the module's auth operation locally. The driver
        module itself is resolved server-side, against this API's own
        module cache, and its auth operation file's raw content is
        delivered directly (authFileContent) — the client needs no local
        hyve.lock/module resolution of its own at all.
      parameters:
        - $ref: "#/components/parameters/ClusterName"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthContextDTO"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: This cluster doesn't use client-side auth — fetch its kubeconfig via GET /api/kubeconfig instead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/secrets:
    get:
      tags: [secrets]
      summary: List secret key names, or (with ?values=true) every key/value
      description: |
        Every secret set via `hyve env secrets` (cluster mode) lives in one
        shared Kubernetes Secret for this hyve-api deployment — not scoped
        per caller/environment. Key names alone are readable by any
        authenticated role; `?values=true` additionally requires the
        `admin` role and returns the full key->value map in one round trip
        (used by cmd/shared's LoadEnvironmentSecrets rather than fetching
        each key individually).
      parameters:
        - name: values
          in: query
          schema:
            type: boolean
          description: Return the full key->value map instead of just key names. Requires the admin role.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { type: string }
                    description: Key names only (default)
                  - type: object
                    additionalProperties: { type: string }
                    description: Full key->value map (?values=true)
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/secrets/{key}:
    get:
      tags: [secrets]
      summary: Get a single secret's value
      description: Requires the `admin` role.
      parameters:
        - $ref: "#/components/parameters/SecretKey"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  key: { type: string }
                  value: { type: string }
                required: [key, value]
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Caller's role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [secrets]
      summary: Add or update a secret
      description: Requires the `admin` role. Creates the shared secrets Secret on first use if it doesn't exist yet.
      parameters:
        - $ref: "#/components/parameters/SecretKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value: { type: string }
              required: [value]
      responses:
        "204":
          description: Set
        "400":
          description: Key doesn't match the required env-var-name pattern (^[A-Za-z_][A-Za-z0-9_]*$), or malformed body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Caller's role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [secrets]
      summary: Remove a secret
      description: Requires the `admin` role. Idempotent — removing a key that isn't set still returns 204.
      parameters:
        - $ref: "#/components/parameters/SecretKey"
      responses:
        "204":
          description: Removed (or already absent)
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Caller's role is not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: accessToken returned by POST /auth/login or /auth/refresh
  parameters:
    ClusterName:
      name: name
      in: path
      required: true
      schema:
        type: string
...
    SecretKey:
      name: key
      in: path
      required: true
      schema:
        type: string
  responses:
...
    Unauthorized:
      description: Missing, malformed, or expired session token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but no matching HyveAccessBinding (or insufficient role)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such object
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AuthContextDTO:
      type: object
      description: |
        Only served for clusters using the default client-side auth method —
        see GET /api/clusters/{name}/auth-context. authFileContent is the
        driver module's resolved auth operation file (auth.yaml/auth.sh),
        resolved server-side against the API's own module cache and
        delivered verbatim — the client writes it to a temp file and runs it
        locally, with no hyve.lock/module resolution of its own required.
      properties:
        driverSource: { type: string }
        driverVersion: { type: string }
        region: { type: string }
        params:
          type: object
          additionalProperties: { type: string }
        driverOutputs:
          type: object
          additionalProperties: { type: string }
        authFileName: { type: string, example: "auth.yaml" }
        authFileContent: { type: string, description: "Raw contents of authFileName, verbatim." }
        tools:
          type: array
          description: The driver module's declared tool requirements (name/description only) — enough for the client to run its usual local-PATH validation before executing authFileContent.
          items:
            type: object
            properties:
              name: { type: string }
              description: { type: string }
            required: [name]
      required: [driverSource, driverVersion, authFileName, authFileContent]
...